Don't try to copy 0 sized files. Fixed incorrect path separator constant.

+++ CoreInteractiveGraph/src/au/gov/asd/tac/constellation/graph/interaction/gui/VisualGraphTopComponent.java
@@ -1111,12 +1111,14 @@ public final class VisualGraphTopComponent extends CloneableTopComponent impleme
                 //        - if backup was not found throw load error
                 final FileObject fileobj = freshGdo.getPrimaryFile();
                 final File srcFile = new File(fileobj.getPath());
-                final String srcfilePath = srcFile.getParent().concat(File.pathSeparator).concat(this.name).concat(".").concat(fileobj.getExt());
+                final String srcfilePath = srcFile.getParent().concat(File.separator).concat(this.name).concat(".").concat(fileobj.getExt());
                 
-                // Create a backup copy of the file before overwriting it. If the backup copy fails, then code will never
-                // get to execute the save, so the actual file should remain intact. If the save fails, the backup file will
-                // already have been written.
-                FileUtils.copyFile(new File(srcfilePath), new File(srcfilePath.concat(BACKUP_EXTENSION)));
+                if(srcFile.exists() && !srcFile.isDirectory() && FileUtils.sizeOf(srcFile) > 0) { 
+                    // Create a backup copy of the file before overwriting it. If the backup copy fails, then code will never
+                    // get to execute the save, so the actual file should remain intact. If the save fails, the backup file will
+                    // already have been written.
+                    FileUtils.copyFile(new File(srcfilePath), new File(srcfilePath.concat(BACKUP_EXTENSION)));
+                }
                 
                 try (OutputStream out = new BufferedOutputStream(freshGdo.getPrimaryFile().getOutputStream())) {
                     // Write the graph.
